Move win32_string_compat.* to os subdirectory.
