+++ avs_core/core/parser/os/win32_string_compat.cpp
